refactor(editor): simplify component hierarchy
