--- packages/editor/src/patterns/Timeline.tsx
@@ -1,3 +1,0 @@
-export const Timeline: React.FC = () => {
-  return null;
-};
